requirements.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
bootstrap-py==1.0.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.2
django-allauth==0.44.0
django-countries==7.2.1
django-crispy==1.8.1
django-crispy-forms==1.11.2
django-environ==0.4.5
django-heroku==0.3.1
django-redis==4.12.1
djangorestframework==3.12.4
docutils==0.17.1
fanstatic==1.1
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.17
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
incremental==21.3.0
Jinja2==3.0.1
js.jquery==3.3.1
MarkupSafe==2.0.1
micropython-fcntl==0.0.3
micropython-ffilib==0.1.2
msgpack==1.0.2
msgpack-python==0.5.6
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
peewee==3.14.4
pep517==0.10.0
pguard==0.4.0
Pillow==8.2.0
pip-tools==6.1.0
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.14.0
PyQt5-sip==12.9.0
pyswip==0.2.10
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rest-condition==1.0.3
service-identity==21.1.0
shutilwhich==1.1.0
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.4
virtualenv==20.4.6
WebOb==1.8.7
whitenoise==5.2.0
zope.interface==5.4.0
